Added cards to game setup activity
