+++ app/build.gradle
@@ -25,6 +25,7 @@ dependencies {
     compile 'com.android.support:appcompat-v7:23.4.0'
     compile 'com.android.support:design:23.4.0'
     compile 'com.android.support:support-v4:23.4.0'
+    compile 'com.android.support:cardview-v7:21.0.+'
     compile 'com.ogaclejapan.smarttablayout:library:1.6.1@aar'
     compile 'com.thebluealliance:spectrum:0.5.0'
 }
