+++ app/src/main/res/layout/activity_game_setup.xml
@@ -27,45 +27,13 @@
                 app:popupTheme="@style/AppTheme.PopupOverlay">
 
             </android.support.v7.widget.Toolbar>
-
         </android.support.design.widget.AppBarLayout>
-
-        <android.support.v4.view.ViewPager
-            android:id="@+id/setupViewPager"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_below="@+id/appbar"
-            app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
-
-        <!--Display Progress-->
-         <com.ogaclejapan.smarttablayout.SmartTabLayout
-            android:id="@+id/smartTab"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:focusable="false"
-            app:stl_clickable="true"
-            app:stl_customTabTextLayoutId="@layout/custom_tab_square"
-            app:stl_dividerColor="@color/transparent"
-            app:stl_dividerThickness="0dp"
-            app:stl_indicatorColor="@color/accent"
-            app:stl_indicatorCornerRadius="4dp"
-            app:stl_indicatorGravity="center"
-            app:stl_indicatorInterpolation="linear"
-            app:stl_indicatorThickness="20dp"
-            app:stl_underlineColor="@color/transparent"
-            app:stl_underlineThickness="0dp"
-            android:layout_above="@+id/createFromSetupButton"
-            android:layout_centerHorizontal="true" />
-
-        <!--Create Game-->
-        <Button
+        <FrameLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:id="@+id/createFromSetupButton"
-            android:text="@string/button_create_game"
-            android:layout_alignParentBottom="true"
-            android:visibility="invisible"/>
+            android:layout_below="@id/appbar"
+            android:id="@+id/setupContainer"/>
+
     </RelativeLayout>
