+++ app/src/main/res/values/strings.xml
@@ -7,8 +7,7 @@
     <string name="action_game_settings">Edit Game Settings</string>
     <string name="action_edit_field_setup">Change Field Setup</string>
     <string name="action_create_new_template">New Template</string>
-    <string name="action_template_save">Save</string>
-    <string name="action_template_save_as">Save As</string>
+    <string name="action_template_create">Create Template</string>
 
     <!--Common-->
     <string name="item_edit">Edit</string>
@@ -89,6 +88,11 @@
     <string name="timepicker_soft_cap_title">Set Soft Cap</string>
     <string name="timepicker_hard_cap_title">Set Hard Cap</string>
     <string name="dialog_validation_failed">Please enter all required fields.</string>
+    <string name="description_winning_score">Game to %1$d</string>
+    <string name="description_no_winning_score">No score limit</string>
+    <string name="description_no_time_caps">No time caps</string>
+    <string name="description_soft_cap">Soft Cap: %1$s</string>
+    <string name="description_hard_cap">Hard Cap: %1$s</string>
 
     <!--Game Defaults-->
     <string name="default_game_name">Ultimate!</string>
@@ -102,4 +106,7 @@
     <string name="dialog_delete_game">Are you sure you want to delete this game?</string>
     <string name="score_list_text">%1$d - %2$d</string>
     <string name="team_list_text">%1$s vs %2$s</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
